move blockchain package to common lib
