+++ apps/worker/src/publisher/ipfs.publisher.ts
@@ -5,10 +5,10 @@ import { ISubmittableResult } from '@polkadot/types/types';
 import { SubmittableExtrinsic } from '@polkadot/api-base/types';
 import { InjectQueue } from '@nestjs/bullmq';
 import { Hash } from '@polkadot/types/interfaces';
-import { BlockchainService } from '../blockchain/blockchain.service';
+import { BlockchainService } from '../../../../libs/common/src/blockchain/blockchain.service';
 import { ConfigService } from '../../../../libs/common/src/config/config.service';
 import { IPublisherJob } from '../interfaces/publisher-job.interface';
-import { createKeys } from '../blockchain/create-keys';
+import { createKeys } from '../../../../libs/common/src/blockchain/create-keys';
 import { IStatusMonitorJob } from '../interfaces/status-monitor.interface';
 import { QueueConstants } from '../../../../libs/common/src';
