+++ libs/common/src/blockchain/blockchain.module.ts
@@ -4,7 +4,7 @@ https://docs.nestjs.com/modules
 
 import { Module } from '@nestjs/common';
 import { BlockchainService } from './blockchain.service';
-import { ConfigModule } from '../../../../libs/common/src/config/config.module';
+import { ConfigModule } from '../config/config.module';
 
 @Module({
   imports: [ConfigModule],
