+++ libs/common/src/blockchain/event-error.ts
